+++ vscode-ext/glyph-middleware.js
@@ -1208,7 +1208,7 @@ class GlyphCompressor {
 
   _createSourceMap() {
     return {
-      version: '1.28.0',
+      version: '1.29.0',
       level: this.level,
       provider: this.provider,
       profile: this.providerProfile,